requirements.txt
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
